requirements.txt
streamlit>=1.20.0
numpy
opencv-python
PyWavelets
Pillow
scipy
matplotlib
